+++ Panther/src/sema_to_pir/SemaToPIRData.cpp
@@ -19,7 +19,7 @@ namespace pcit::panther{
 
 
 	auto SemaToPIRData::createJITInterfaceFuncDecls(pir::Module& module) -> void {
-		const auto create_func_decl = [&](std::string_view name, evo::SmallVector<pir::Parameter>&& params){
+		const auto create_return_func_decl = [&](std::string_view name, evo::SmallVector<pir::Parameter>&& params){
 			return module.createExternalFunction(
 				std::string(name),
 				std::move(params),
@@ -30,7 +30,7 @@ namespace pcit::panther{
 		};
 
 
-		this->jit_interface_funcs.return_generic_int = create_func_decl(
+		this->jit_interface_funcs.return_generic_int = create_return_func_decl(
 			"PIR.JIT.return_generic_int",
 			{
 				pir::Parameter("target", module.createPtrType()),
@@ -39,39 +39,74 @@ namespace pcit::panther{
 			}
 		);
 
-		this->jit_interface_funcs.return_generic_bool = create_func_decl(
+		this->jit_interface_funcs.return_generic_bool = create_return_func_decl(
 			"PIR.JIT.return_generic_bool",
 			{pir::Parameter("target", module.createPtrType()), pir::Parameter("value", module.createBoolType())}
 		);
 
-		this->jit_interface_funcs.return_generic_f16 = create_func_decl(
+		this->jit_interface_funcs.return_generic_f16 = create_return_func_decl(
 			"PIR.JIT.return_generic_f16",
 			{pir::Parameter("target", module.createPtrType()), pir::Parameter("value", module.createPtrType())}
 		);
-		this->jit_interface_funcs.return_generic_bf16 = create_func_decl(
+		this->jit_interface_funcs.return_generic_bf16 = create_return_func_decl(
 			"PIR.JIT.return_generic_bf16",
 			{pir::Parameter("target", module.createPtrType()), pir::Parameter("value", module.createPtrType())}
 		);
-		this->jit_interface_funcs.return_generic_f32 = create_func_decl(
+		this->jit_interface_funcs.return_generic_f32 = create_return_func_decl(
 			"PIR.JIT.return_generic_f32",
 			{pir::Parameter("target", module.createPtrType()), pir::Parameter("value", module.createFloatType(32))}
 		);
-		this->jit_interface_funcs.return_generic_f64 = create_func_decl(
+		this->jit_interface_funcs.return_generic_f64 = create_return_func_decl(
 			"PIR.JIT.return_generic_f64",
 			{pir::Parameter("target", module.createPtrType()), pir::Parameter("value", module.createFloatType(64))}
 		);
-		this->jit_interface_funcs.return_generic_f80 = create_func_decl(
+		this->jit_interface_funcs.return_generic_f80 = create_return_func_decl(
 			"PIR.JIT.return_generic_f80",
 			{pir::Parameter("target", module.createPtrType()), pir::Parameter("value", module.createPtrType())}
 		);
-		this->jit_interface_funcs.return_generic_f128 = create_func_decl(
+		this->jit_interface_funcs.return_generic_f128 = create_return_func_decl(
 			"PIR.JIT.return_generic_f128",
 			{pir::Parameter("target", module.createPtrType()), pir::Parameter("value", module.createPtrType())}
 		);
 
-		this->jit_interface_funcs.return_generic_char = create_func_decl(
+		this->jit_interface_funcs.return_generic_char = create_return_func_decl(
 			"PIR.JIT.return_generic_char",
 			{pir::Parameter("target", module.createPtrType()), pir::Parameter("value", module.createIntegerType(8))}
+		);
+
+
+
+
+		this->jit_interface_funcs.get_generic_int = module.createExternalFunction(
+			"PIR.JIT.get_generic_int",
+			{pir::Parameter("source", module.createPtrType()), pir::Parameter("value", module.createPtrType())},
+			pir::CallingConvention::C,
+			pir::Linkage::EXTERNAL,
+			module.createVoidType()
+		);
+
+		this->jit_interface_funcs.get_generic_bool = module.createExternalFunction(
+			"PIR.JIT.get_generic_bool",
+			{pir::Parameter("source", module.createPtrType())},
+			pir::CallingConvention::C,
+			pir::Linkage::EXTERNAL,
+			module.createBoolType()
+		);
+
+		this->jit_interface_funcs.get_generic_float = module.createExternalFunction(
+			"PIR.JIT.get_generic_float",
+			{pir::Parameter("source", module.createPtrType()), pir::Parameter("value", module.createPtrType())},
+			pir::CallingConvention::C,
+			pir::Linkage::EXTERNAL,
+			module.createVoidType()
+		);
+
+		this->jit_interface_funcs.get_generic_char = module.createExternalFunction(
+			"PIR.JIT.get_generic_char",
+			{pir::Parameter("source", module.createPtrType())},
+			pir::CallingConvention::C,
+			pir::Linkage::EXTERNAL,
+			module.createIntegerType(8)
 		);
 	}